Move shared errors and split data_converters
