+++ crates/sdk/src/error.rs
@@ -1,0 +1,6 @@
+//! Shared SDK error re-exports.
+
+pub use temporalio_common::error::{
+    ActivityExecutionError, ApplicationFailure, ChildWorkflowExecutionError,
+    ChildWorkflowSignalError,
+};
